+++ src/body/params.ts
@@ -5,6 +5,13 @@ import {URLSearchParams} from 'node:url';
  */
 export class Params extends URLSearchParams {
   /**
+   * Check if parameters are empty.
+   */
+  get isEmpty(): boolean {
+    return [...this].length === 0;
+  }
+
+  /**
    * Convert parameters into a plain object, useful for validation.
    */
   toObject<T extends Record<string, any>>(): T {